requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
deep-translator==1.10.1
docxcompose==1.4.0
docxtpl==0.16.5
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
maxdmg-back==0.0.4
maxdmg-docgen==0.0.3
maxdmg-resource==0.0.2
multidict==6.0.4
pyairtable==1.4.0
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
